+++ lib/modules/platform/github/issue.ts
@@ -1,7 +1,6 @@
 import { DateTime } from 'luxon';
 import { z } from 'zod';
 import { getCache } from '../../../util/cache/repository';
-import type { GithubIssue as Issue } from './types';
 
 const GithubIssueBase = z.object({
   number: z.number(),
@@ -10,9 +9,17 @@ const GithubIssueBase = z.object({
   body: z.string(),
 });
 
+export interface GithubIssue {
+  body: string;
+  number: number;
+  state: string;
+  title: string;
+  lastModified: string;
+}
+
 const GithubGraphqlIssue = GithubIssueBase.extend({
   updatedAt: z.string(),
-}).transform((issue): Issue => {
+}).transform((issue): GithubIssue => {
   const lastModified = issue.updatedAt;
   const { number, state, title, body } = issue;
   return { number, state, title, body, lastModified };
@@ -20,14 +27,13 @@ const GithubGraphqlIssue = GithubIssueBase.extend({
 
 const GithubRestIssue = GithubIssueBase.extend({
   updated_at: z.string(),
-}).transform((issue): Issue => {
+}).transform((issue): GithubIssue => {
   const lastModified = issue.updated_at;
   const { number, state, title, body } = issue;
   return { number, state, title, body, lastModified };
 });
 
 export const GithubIssue = z.union([GithubGraphqlIssue, GithubRestIssue]);
-export type GithubIssue = z.infer<typeof GithubIssue>;
 
 type CacheData = Record<number, GithubIssue>;
